+++ manifest/clients/cohere.py
@@ -44,7 +44,7 @@ class CohereClient(Client):
             connection_str: connection string.
             client_args: client arguments.
         """
-        self.api_key = os.environ.get("COHERE_API_KEY", connection_str)
+        self.api_key = connection_str or os.environ.get("COHERE_API_KEY")
         if self.api_key is None:
             raise ValueError(
                 "Cohere API key not set. Set COHERE_API_KEY environment "
@@ -93,7 +93,7 @@ class CohereClient(Client):
         """
         return {"model_name": self.NAME, "engine": getattr(self, "engine")}
 
-    def validate_response(self, response: Dict, request: Dict) -> Dict[str, Any]:
+    def postprocess_response(self, response: Dict, request: Dict) -> Dict[str, Any]:
         """
         Format response to dict.