+++ manifest/clients/openai.py
@@ -60,7 +60,7 @@ class OpenAIClient(Client):
             connection_str: connection string.
             client_args: client arguments.
         """
-        self.api_key = os.environ.get("OPENAI_API_KEY", connection_str)
+        self.api_key = connection_str or os.environ.get("OPENAI_API_KEY")
         if self.api_key is None:
             raise ValueError(
                 "OpenAI API key not set. Set OPENAI_API_KEY environment "
@@ -107,7 +107,7 @@ class OpenAIClient(Client):
         """
         return {"model_name": self.NAME, "engine": getattr(self, "engine")}
 
-    def validate_response(self, response: Dict, request: Dict) -> Dict[str, Any]:
+    def postprocess_response(self, response: Dict, request: Dict) -> Dict[str, Any]:
         """
         Validate response as dict.
 
@@ -118,7 +118,7 @@ class OpenAIClient(Client):
         Return:
             response as dict
         """
-        validated_response = super().validate_response(response, request)
+        validated_response = super().postprocess_response(response, request)
         # Handle logprobs
         for choice in validated_response["choices"]:
             if "logprobs" in choice: