+++ manifest/manifest.py
@@ -474,6 +474,7 @@ class Manifest:
         stop_token: Optional[str] = None,
         return_response: bool = False,
         chunk_size: int = -1,
+        verbose: bool = False,
         **kwargs: Any,
     ) -> Union[List[str], List[np.ndarray], Response]:
         """
@@ -500,6 +501,7 @@ class Manifest:
                 For a single manifest client, there is no impact to
                 setting chunk_size. For a client pool, chunk_size
                 can be used to distribute the load across the clients.
+            verbose: whether to print progress of async tasks.
 
         Returns:
             response from prompt.
@@ -530,6 +532,7 @@ class Manifest:
                         prompts=chunk,
                         client=client,
                         overwrite_cache=overwrite_cache,
+                        verbose=verbose,
                         **kwargs,
                     )
                 )
@@ -553,6 +556,7 @@ class Manifest:
         prompts: List[str],
         client: Client,
         overwrite_cache: bool = False,
+        verbose: bool = False,
         **kwargs: Any,
     ) -> Response:
         """
@@ -562,6 +566,7 @@ class Manifest:
             prompts: prompts to run.
             client: client to run.
             overwrite_cache: whether to overwrite cache.
+            verbose: whether to print progress of async tasks.
 
         Returns:
             response from prompt.
@@ -579,7 +584,7 @@ class Manifest:
         # If not None value or empty list - run new request
         if request_params.prompt:
             self.client_pool.start_timer()
-            response = await client.arun_batch_request(request_params)
+            response = await client.arun_batch_request(request_params, verbose=verbose)
             self.client_pool.end_timer()
         else:
             # Nothing to run